% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4-ProjectionClass.R
\docType{class}
\name{Projection-class}
\alias{Projection-class}
\alias{Projection-class}
\alias{Projection}
\alias{vec}
\alias{vec,Projection-method}
\alias{bounds}
\alias{bounds,Projection-method}
\alias{mat}
\alias{mat,Projection-method}
\alias{Aseq}
\alias{Aseq,Projection-method}
\alias{projtype}
\alias{projtype,Projection-method}
\alias{vectype}
\alias{vectype,Projection-method}
\alias{nmat}
\alias{nmat,Projection-method}
\alias{show,Projection-method}
\alias{plot,Projection-method}
\title{S4 class describing a 'Projection' object}
\usage{
vec(object)

\S4method{vec}{Projection}(object)

bounds(object)

\S4method{bounds}{Projection}(object)

mat(object, ...)

\S4method{mat}{Projection}(object, return = "simple")

Aseq(object)

\S4method{Aseq}{Projection}(object)

projtype(object)

\S4method{projtype}{Projection}(object)

vectype(object)

\S4method{vectype}{Projection}(object)

nmat(object)

\S4method{nmat}{Projection}(object)

\S4method{show}{Projection}(object)

\S4method{plot}{Projection}(x, bounds = FALSE, bounds.args = NULL,
  labs = TRUE, plottype = "lines", ybreaks = 20, shadelevels = 100, ...)
}
\arguments{
\item{object}{an object of class "Projection" generated using \code{\link{project}}}

\item{...}{arguments to be passed to methods: see \code{\link{par}} and
\code{\link{plot}}.}

\item{return}{either "simple", "list", or "array": used for accessing the 'mat'
slot from a Projection object. Note that only list or array can be used for 
stochastic projections, which have more than one matrix.
Access population vectors (slot "vec")}

\item{x}{a Projection object created using \code{\link{project}}}

\item{bounds}{logical: indicates whether to plot the bounds on population density.}

\item{bounds.args}{A list of graphical parameters for plotting the bounds if
\code{bounds=T}. The name of each list element indicates the name of the argument.
Could include, e.g. \code{list(lwd=2,lty=3,col="darkred")}.}

\item{labs}{logical: if \code{TRUE}, the plot includes more than one projection
and \code{plottype="lines"}, then lines are automatically labelled according to 
the names contained in the 'projection' object.}

\item{plottype}{for projections generated from dirichlet draws (see 
\code{\link{project}}), \code{plottype} has two options. \code{"lines"}
will plot each projection as a separate line. \code{"shady"} will 
plot shaded contours showing the probabilities of population 
densities over over time, calculated across the set of projections from 
dirichlet draws. By default this shaded plot is a gradient of black to
white (with black representing higher probabilities), but this can be
overridden by using the 'col' argument (see examples).}

\item{ybreaks}{if \code{plottype="shady"}, gives the number of breaks on
the y axis for generating the grid for the shade plot. A larger number of
breaks means a finer resolution grid for the shading.}

\item{shadelevels}{if \code{plottype="shady"} and a palette of colours
is not specified using 'col', then \code{shadelevels} gives the number of
colour/shading levels to use when generating the black and white shade plot. 
A larger number of levels means a finer resolution on the shade
plot of population density (see examples).}
}
\description{
Projection objects are created using the \code{\link{project}} function. 
The slots of Projection objects are described with reference to arguments 
passed to the \code{project} function.

Access bounds on population dynamics (slot "bounds")

Access population vectors (slot "vec")

Access projection matrix sequence (slot "Aseq")

Access projection type (slot "projtype")

Access initial vector type (slot "vectype")

Access number of projection matrices

Plot method for Projection class
}
\section{Slots}{

\describe{
\item{\code{.Data}}{One or more time series of population sizes. 
'Projection' objects inherit from a standard array, and can be treated as 
such. Therefore, if \code{vector} is specified, the 'Projection' object will 
behave as: 
\itemize{
 \item if a single \code{vector} is given, a numeric vector of population sizes 
 of length \code{time+1}
 \item if multiple \code{vector}s are given, a numeric matrix of population 
 projections where each column represents a single population projection and 
 is of length \code{time+1}
 \item if \code{vector="n"}, a numeric matrix of population projections where each column 
 represents a single stage-biased projection and is of length \code{time+1}.
 \item if \code{vector="diri"}, a numeric matrix of population projections where each 
 column represents projection of a single vector draw and each column is of 
 length \code{time+1}.
}}

\item{\code{vec}}{Age- or stage-based population vectors. \code{vec} 
will be:
\itemize{
 \item If a single \code{vector} is specified, a numeric matrix of demographic 
 vectors from projection of \code{vector} through \code{A}. Each column 
 represents the densities of one life (st)age in the projection.
 \item If multiple \code{vector}s are specified, a three-dimensional array of
 demographic vectors from projection of the set of initial vectors through
 \code{A}. The first dimension represents time (and is therefore equal to 
 \code{time+1}). The second dimension represents the densities of 
 each stage (and is therefore equal to the dimension of \code{A}). 
 The third dimension represents each individual projection (and is 
 therefore equal to the number of initial vectors given).
 \item If \code{vector="n"}, a three-dimensional array of demographic vectors from
 projection of the set of stage-biased vectors through \code{A}. The first 
 dimension represents time (and is therefore equal to \code{time+1}). The 
 second dimension represents the densities of each stage (and 
 is therefore equal to the dimension of \code{A}). The third 
 dimension represents each individual stage-biased projection (and is 
 therefore also equal to the dimension of \code{A}).
 \item If\code{vector="diri"}, a three-dimensional array of demographic vectors from
 projection of the dirichlet vector draws projected through \code{A}. The first 
 dimension represents time (and is therefore equal to \code{time+1}). The second 
 dimension represents the densities of each stage (and 
 is therefore equal to the dimension of \code{A}). The third 
 dimension represents projection of each population draw (and is therefore equal
 to \code{draws}).
}\cr

Some examples for understanding the structure of 3D arrays returned when 
\code{return.vec=TRUE}: when projecting a 3 by 3 matrix for >10 time intervals, 
element [11,3,2] represents the density of stage 3 at time 10 
for either vector 2 (multiple vectors), stage-bias 2 (\code{vector="n"}) or draw 2 
(\code{vector="diri"}); note that because element 1 represents t=0, then t=10 
is found at element 11. The vector [,3,2] represents the time series of densities 
of stage 3 in the projection of vector 2 / stage-bias 2 / draw 2. The matrix [,,2] 
represents the time series of all stages in the projection of vector 2 / stage-bias 
2 / draw 2.\cr\cr

Note that the projections inherit the labelling from \code{A} and \code{vector}, if
it exists. Both stage and vector names are taken from the COLUMN names of \code{A} 
and \code{vector} respectively. These may be useful for selecting from the
\code{projection} object, and are used when labelling plots of Projection 
objects containing multiple population projections.\cr\cr

Set \code{return.vec = FALSE} when calling \code{project} to prevent population
vectors from being saved: in this case, \code{vec} is equal to 
\code{numeric(0)}. This may be necessary when projecting large numbers of
vectors, as is the case when \code{vector = "diri"}.}

\item{\code{bounds}}{The bounds on population dynamics (only for deterministic 
projections). These represent the maximum and minimum population sizes 
achieveable at each time interval of the projection. \code{bounds} is a 
matrix with 2 columns (lower and upper bounds, in that order), and the 
number of rows is equal to \code{time + 1}.}

\item{\code{mat}}{The matrix/matrices used in the population projection. In their 
raw form \code{mat} is always a three-dimensional array, where the third 
dimension is used to index the different matrices. However, by using the 
\code{mat()} accessor function below, it is possible to choose different ways
of representing the matrices (matrix, list, array).}

\item{\code{Aseq}}{The sequence of matrices used in the projection. For deterministic
projections (where there is only 1 matrix) this will always be \code{rep(1, time)}.
For stochastic projections (with more than 1 matrix), if \code{Aseq} is given 
to \code{project} as a numeric or character vector then this slot will take 
that value. If a matrix describing a random markov process is passed, the 
\code{Aseq} slot will be a single random chain.}

\item{\code{projtype}}{The type of projection. Either "deterministic" (single matrix; 
time-invariant), or "stochastic" (multiple matrices; time-varying).}

\item{\code{vectype}}{The type of vector passed to \code{project}. May be "single" 
(one vector; one population projection), "multiple" (more than one vector; 
several population projections), "bias" (stage-biased vectors; 
\code{vector = "n"}), or "diri" (vectors drawn from the dirichlet 
distribution; \code{vector = "diri"}).}
}}

\section{Generics & Methods}{
 
Accessing slots and other information from a Projection object
}

\section{Display Methods}{

Printing and plotting Projection objects
Print method for Projection class
}

